moved TaskViewerState to doubtfire.states.ts
